Execute print user script in page content world (#238)
The print user script needs to execute in the page content world so that
it mutates the globals used by the page which the user is interacting
with. The existing setup runs scripts under the default content world
(which is what you generally want), but this script is a special case as
we want to overwrite a global function on the window object.