Attempt to fix a bug when external URL is searched for
Scenario:
1. Enter `mailto:test@example.com`, submit: confirmation popover appears
2. Without reloading the page, enter a search phrase in the address bar
3. The search is performed for `mailto:test@example.com`

This is caused by PermissionAuthorizationQuery deinitializer
which calls decisionHandler that ends up requesting a search
for the external URL. The query is deinitialized when the new search
phrase is submitted and the sequence of events is following:
1. make search URL for the search phrase
2. make search URL for the external URL from a deinitialized PermissionAuthorizationQuery
The latter overrides the search phrase and that's how the bug is
observed.

I updated PermissionModel.queryAuthorization to not call its decision
handler at all when decision from query is `.deinitialized`.