Fix handling external schemes in address bar (#493)
Task/Issue URL: https://app.asana.com/0/inbox/1201621708108816/1202055192971137/1202055303685720/f

Description:
The original issue is about define:foo being recognized as an external URL (with define: scheme) and not triggering a search, but instead opening a blank page. I realized that other external URLs don't work either when typed directly into the address bar. I figured out that it's because of a recently modified code in Tab.swift which allows for handling external URLs (optionally converting them to search URLs if they can't be handled by NSWorkspace) only when the source of a navigation is the main frame. However, when user enters a URL into the address bar, the source of navigation is not main frame either, so I handled it by checking that navigation type is .other (which covers manually input URLs).