Fix animation hitching. (#518)
Task/Issue URL: https://app.asana.com/0/1200835453786799/1202102930180367/f
Tech Design URL:
CC:

Description:

This PR improves the animation when displaying the Privacy Dashboard.

It does this by avoiding running the resize animation block during the popover's own animation. The approaches uses a new boolean to track whether or not to run the animation, but it would be nice to not even have to rely on that if anyone has a suggestion for a better approach.