Autofill scrolling fix (#538)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1202037446274716/f
Tech Design URL:
CC:

Description:

This PR removes the GeometryReader logic that was used to do some scroll-dependent opacity changes to a divider. This opacity feature wasn't requested in the design spec, instead it was a nice-to-have that I added and the design team were happy to keep.

However, we've had an issue recently where Gabe has had Autofill get stuck when scrolling. I was able to reproduce this one (and never since then), and when it was happening I was getting hundreds of errors in the Xcode console about the GeometryReader.

To be safe, the code is being removed. Gabe has had a build with this change since last week and has said it's working great, so let's put it into production and see if it holds up.