Add data import failure pixels (#552)
Task/Issue URL: https://app.asana.com/0/0/1202158855860055/f
Tech Design URL:
CC: @brindy

Description:

This PR adds a new pixel which reports the import source, import type, and error message when data import fails.

The format and parameters of this pixel are found under the linked Asana task.

The pixel from this PR will be used to drill down into where the importer is failing and why next week, and translate into some fixes. One thing we're missing is the actual versions of the files we're reading from, which will be added in a follow-up task later.