Update the Fireproof checkmark in the Save Credentials view controller (#555)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1202118895326543/f
Tech Design URL:
CC:

Description:

This PR updates the Save Credentials VC to accurately reflect whether a domain is already fireproof, and to change that in the database if the checkmark changes. That is, if a domain is already fireproof and you un-check that box before saving credentials, it will now remove it from the Fireproof store.

This PR has also taken a pass over the debug menu, moving all of the "Reset *" options into their own "Reset Data" submenu, and adding a new "UI Triggers" submenu for triggering pieces of the UI which can be annoying to iterate on if you need to run through a bunch of steps on a website. There's only one option in this menu for now, which is to trigger the Save Credentials VC.

Note: I went back and forth on a full refactor of this popover, but have since been given another task which requires making larger changes to this UI, so I've decided to just open this PR now with the simplest changes possible in order to get this fixed.