Improve build phase times during compilation (#607)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1202410142344161/f
Tech Design URL:
CC:

Description:

This PR improves the app's compile times by replacing our Check Filename Headers script with SwiftLint. The existing script was taking a while, sometimes over 10 seconds, even on an M1 Pro.

As a part of this change, I fixed existing header comments that were not aligned with our default. We had a number of headers that included the target name, but most didn't, so I removed target name from any existing comments that have it. I prefer it this way, but have no problem if we want to enforce the target name instead – just let me know and I'll update the PR.

Note: I implemented this change on Xcode 14. I haven't been able to verify it against Xcode 13, as 13 is now crashing on launch every single time. I'll try to fix this, but if you can review this branch on Xcode 13 that would be amazing. 🙏