Add support for the new interactionState property (#653)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1202654408128835/f
Tech Design URL:
CC:

Description:

This PR adds support for the interactionState property added in macOS 12. This is the new public API used for saving and restoring state.

The existing code is still in place, and is used to migrate users to the new format.