Fix focus bug with pinned tabs (#667)
Task/Issue URL: https://app.asana.com/0/0/1202668155654614/f

Description:
TL;DR: This patch removes unnecessary calls to showTabContent.

The fix for while background flashing introduced delaying a call to showTabContent
until any of webViewDidCommit, webViewDidFail or webViewDidReceiveChallenge is called
(with this code specifically).
This was done to ensure that tab content is only updated when the website actually starts loading
(or fails or requires authentication).
Only one of these events is needed to switch to the new tab content, so the merge publisher
should be limited to a single event. It appears that when navigating to websites behind DDG SSO,
such as Asana, webViewDidReceiveChallenge is called also after webViewDidCommit
(sometimes multiple times), which seems to occasionally be breaking keyboard input.