Update selectedTabViewModel only when it actually changes (#681)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1202668155654614/f

Description:
TL;DR: This patch limits the number of events published from
TabCollectionViewModel.selectedTabViewModel to reduce number
of calls to BrowserTabViewController.showTabContent.

Before this fix, when you navigated between tabs using cmd+[], showTabContent
would be called once for every unpinned tab, but twice for every pinned tab.
This is because two observable streams are activated after TabCollectionViewModel
updates its selectionIndex after a call to selectNext/selectPrevious:

* TabCollectionViewModel updates PinnedTabsViewModel's selectedItem via
  a subscription in TabBarViewController,
* PinnedTabsViewModel tries to update TabCollectionViewModel's selectionIndex again
  (to the same pinned tab index) via a subscription in TabBarViewController.

Both subscriptions are currently required, because the latter is needed to support clicking
pinned tabs with a mouse, when the action originates in PinnedTabView that knows nothing about
TabCollectionViewModel, so it calls PinnedTabsViewModel. By updating selectedTabViewModel
only when the new value is a different instance we can ensure that subscribers are not getting
duplicate events (it can't be easily solved with removingDuplicates, because TabViewModel
is a reference type, plus it's better to solve the publisher itself).