Use LazyVStack to improve new tab page performance. (#699)
This PR updates the new tab page to use a LazyVStack on macOS 11 and up. This dramatically increases performance for users who have 1000+ bookmarks.

To make this easier, I pulled a couple small custom buttons out into their own views, and then did the same with the main VStack code.