Force callers to refresh even when the save fails. (#701)
It's possible for the bookmarks store to return an error even when the actual save call succeeded, as Core Data can return errors unrelated to the object you're saving. In this case, the caller will not refresh itself, even despite the data actually having successfully changed under the hood.

This PR updates the failure completion handlers to trigger a refresh regardless. There's more work to do here to track when and how failures happen at all, but this will solve the problem that Alex was seeing.

The issue that Alex is seeing is already repaired when you launch the app, but it's not repaired if an error is encountered at runtime. I'd like to fix that, but want to take more time to do it correctly, so I'll prepare another PR next week – this PR will be enough to get around his issue.