Update the lockscreen function to return a more accurate value. (#730)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1203070153945517/f
Tech Design URL:
CC:

Description:

This PR updates the return value of the Lock Screen function, to ensure that the UI isn't disabled in cases when it shouldn't be. This state should never be able to happen, but this change is made speculatively to avoid a bug where the browser can get into a weird state.