Update the bookmark import strategy, and import favicons (#708)
Task/Issue URL: https://app.asana.com/0/0/1202872661718609/f
Tech Design URL:
CC:

Description:

This PR updates the way we import bookmarks to meet the following criteria:

• Improve the import logic to result in more expected bookmarks bar and favorites grid:
    • If the bookmarks folder is empty, do the following:
        • Find the browser’s "special" folder. Safari → Favorites; Firefox → Bookmarks Toolbar; Chrome → Bookmarks Bar.
        • Add the contents of the special folder, including sub folders (but not the special folder container itself) to the root bookmarks folder.
        • Favorite any individual bookmarks now at the root of this special folder.
        • Import any other items (folders or bookmarks), outside of this special folder, exactly as they are (including top level folders) into the root bookmarks folder.
    • If the bookmarks folder is not empty, create a new folder called "Imported from <Browser Name>" and import the exact folder structure as-is into this folder. Favorite nothing.

It additionally adds the ability for the browser to read the favicon databases for all browsers that we import from. To facilitate Safari import, the import flow had to be updated to ask the user to grant us access to their entire Safari data directory, not just the Bookmarks file. This has already gone through privacy triage.