Bookmarks store structure fixes (#747)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1203078052409045/f
Tech Design URL:
CC:

Description:

This PR makes a pretty horrid attempt at fixing an issue where the bookmarks store can somehow end up in an invalid state. It's possible for it to acquire bookmarks that have no data attached to them - no URL, no title. The database has a strict non-optional constraint for the title value, so when this happens and you try to save something to the DB, it fails.

This PR adds a step at launch that scans for these bookmarks and removes them. Tim encountered this issue with the production browser, and this branch fixed the issue for him with no data loss.