Update autoconsent to 3.0.2 to fix issue with Sourcepoint CMP (#772)
Task/Issue URL: https://app.asana.com/0/1201844467387842/1203144940444659/f

Description:

This is a fix to the issue mentioned in the Asana task, i.e. a wrongful detection of a CMP on derstandard.at which causes the UI dialogue for the first-time autoconsent setup to show up but will then not produce any visible effect as the CMP can't be handled.