Bug: Privacy Dashboard inaccessible after being dismissed during rules recompilation (#785)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1202821817932095/f

Description:

We have some internal state, pendingUpdates, where we recorded completion tokens
following a click on 'toggle protections' in the dashboard.

We added an entry to this dictionary when the button is clicked -> this is how we determined
isPendingUpdates. A none-empty dictionary means we've performed an action, but we haven't
received confirmation that it was successful.

When we observe an event from userContentBlockingAssets, we look at our internal state
and remove the entry if it matches: this works fine if you stay on the same tab.

There's a bug though: the dashboard's webview can be dismissed before the completion event
is observed if you switch tabs quickly enough -> this cancels all subscriptions and we don't get
a chance to remove the completion token from our own internal state (because we never 'see'
the event telling us it was successful ).

We end up in a  zombie state  with a token being stored internally (which causes
isPendingUpdates to always be true) and therefore we don't allow the dashboard to be opened
on subsequent clicks, until the app is restarted.

This change ties rules recompilation subscription to the lifetime of Privacy Dashboard
(and not its current tabViewModel), so that rules recompilation completion can be properly
handled even when PD popover has been dismissed).