Fix burning Duck Player history entries (#824)
Task/Issue URL: https://app.asana.com/0/0/1203322325220637/f

Description:
Duck Player puts "Duck Player" in place of the domain on the New Tab Page Privacy Feed.
It was stored in RecentlyVisitedSiteModel.domain while setting RecentlyVisitedSiteModel.isRealDomain to false.
As a consequence, the domain string (equal to "Duck Player") would be passed to the burn function, causing
Duck Player entries to persist burning.

This change updates RecentlyVisitedSiteModel to always store the actual domain in domain, and adds
private domainPlaceholder (that defaults to nil) and public domainToDisplay computed property that
takes domainPlaceholder if it exists, or falls back to domain. This way, the proper functionality of burning
from the Privacy Feed is restored (and any new functionality based on RecentlyVisitedSiteModel.domain
should work properly).