release popover memory (#816)
This PR focusses on the popovers from the NavigationBarViewController which contain two of the biggest offenders: bookmarks and autofill. I extracted all the popover related code from the VC and put it in a dedicated class, then just call it from the VC.

Note that these do not "leak", they where just retaining the memory they use unnecessarily by keeping a strong reference to the popover once it was created so that it can be re-used without being recreated.

The privacy dashboard was a candidate, but flickers when it is recreated on the fly and that seemed too big to deal with for this quick maintenance task. There are also over popovers that were not covered in this PR, but they do not hold on to a lot of data and are either simple input forms or for messaging).