Update the history save failure pixel. (#846)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1203386050668201/f
Tech Design URL:
CC:

Description:

This PR updates the pixel fired when failing to save history. The pixel it was using was historyCleanEntriesFailed, but this is in use elsewhere, and it turns out there was an unused historySaveFailed pixel - it seems like this was a typo.