Support for reordering Favorites (#845)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1202117615001669/f
Tech Design URL: https://app.asana.com/0/481882893211075/1203352064964054/f

Description:
This change updates how Favorites are stored in Core Data. A dedicated Favorites Folder is now created.
This folder is outside of the Bookmarks tree. Individual bookmarks are marked as favorite by creating
a relationship to this folder (the relationship is called favoritesFolder and is separate from parentFolder).
The relationship on the Favorites Folder side is ordered, which is how favorites order is persisted.

BookmarkStore and BookmarkManager get moveFavorites API which is sibling to move(objectUUIDs:) but
works on favorites relationship (without affecting Bookmarks folder tree).

Reordering is available:
* from Bookmarks Management on all operating systems
* from the Home Page favorites view on macOS Big Sur and newer (the view uses LazyVGrid introduced in BigSur).

Heavyweight database migration was added to update isFavorite Core Data attribute to the relationship.