Allow blank Firefox bookmarks to be imported (#859)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1203404505761428/f
Tech Design URL:
CC:

Description:

This PR allows Firefox bookmarks with no title to be imported.

As a part of testing this, I also fixed an error value which was duplicated in the importer. Now the failed import will distinguish between imports where the logins file was not found, and where the login file was unreadable. This first case is valid, and can happen if you have never saved a login in Firefox.