Support JS identity pixel (#871)
Task/Issue URL: https://app.asana.com/0/0/1203478208847846/f
Tech Design URL:

Description:

This is the macOS counterpart of duckduckgo/BrowserServicesKit#190.

It adds support for the JS layer to send pixels. Currently, the only pixel being sent is one for when identities are filled. This is done because the JS layer manages the autofilling of identities on its own, compared to passwords and cards which go through the native layer when they are filled.