Replace problematic History BatchDeletes with fetch + deletes (#873)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1203476568967130/
Description:
We noticed a fairly high rate of history_save_failed pixels over the last while. On updating the HistoryStore tests and writing some to try saving after performing a few of the other CoreData operations, it seems that the batch deletes are causing inconsistencies. Generally it's somewhat discouraged to use batch deletes on entities with relationships as this will always be a risk. To fix the tests (and hopefully bring the rate of these errors down), I've replaced the problematic batch deletes in the HistoryStore with simpler fetch + delete operations.

* Update HistoryStoreTests to use real DB

* Use UUID not classname for store name

this will accumulate stores... but maybe that's fine for now

* Throw if the teardown save fails

* Add a test to check that the visits

* Update tests

* Don't use batch delete for cleaning

* Replace batch delete from removeVisits

* Remove extra clean old test

* Extract save history with visits code

* Add additional test for save after removeEntries

* Remove entire temp directory after test

* Free up more memory at end of tests

* Ensure pixels are fired on failure

* Fire Pixel on save initial fetch failure

* Update removeEntries to not use batch delete

* Actually delete the visits

* Improve / write more deletion tests

* Write a test for cascade deletion

* Tidy up a bit

* Add visits count test

* Fix issue with inconsistent visits count

* Drop numberOfTotalVisits from History.HistoryEntry

* Actually commit model update

* Revert visit count changes for now

* Add missing errors to Pixels

* Move initialisation of context into init

* Fix timing issue with deleting visits + entries

Co-authored-by: Bartek Waresiak <bartek@duckduckgo.com>