Revert removal of History context lazy load (#902)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1203560842625353/f
Description:
There's been a rise in database.make.database.error Pixels 0.31.2. This could be due to a change in the timing of when the HistoryStore database gets initialised (it was previously lazy loaded, but it got moved into the initialiser).
It's not crazy dramatic (about 5-10 per day) but we should fix it soon as the consequences are pretty bad (History not working).