Fix the testing steps list in the PR template. (#910)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1203591896442256/f
Tech Design URL:
CC:

Description:

This PR updates the testing steps section of the template.

The reason is this: when you are editing a list in GitHub and hit return, it will automatically increment the list and put you on a new line. Also, when you have a list that starts with 1. on each line, GitHub will calculate its actual offset automatically.

However, our list starts with two 1. entries, meaning that when you hit return on the second line you get this:

1.
1.
2.
3.
etc.

This PR removes the second 1. entry so that the list indices are accurate. Note: GitHub still displays the list correctly when formatted above, so this PR is almost inconsequential – it's just that I keep manually fixing the list when this happens, and I'd rather just fix it for good. 🙂