Fix add bookmark / folder modal in fullscreen (#930)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1202138144256919/f
Description:

Note: This is no longer a problem on Ventura. But this fixes the issue on earlier OS versions

When adding a bookmark via the three dots menu on the top right, while having the browser in full screen, the popup window for adding that bookmark goes fullscreen as well. After finishing, the full screen window stays, but black.

This happens for every modally presented ViewController, not just from the three dot menu.

In most cases, this was fixable by just presenting as a sheet, but for the Add Bookmark / Folder modal from the Bookmarks popover, it a.) needed to be presented from the main window (not the popover's) and b.) needs a slightly hacky solution to prevent the popover from being closed (as AppKit doesn't really like it when you present modals from popovers nor present a modal when there's a popover being displayed).

* Helper extension to begin sheet from main window

* Fix fullscreen bug for bookmark / folder add

* Prevent popover from closing on presenting sheet

* Linting

* Set foldermodalVC delegate

* SwiftLint