Allow importing from Firefox when it doesn’t have logins. (#945)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1203407584243670/f
Tech Design URL:
CC:

Description:

This PR improves Firefox login import in a few ways:

The profile code was only checking for the presence of a logins file, and considering the profile empty if one was not found. This meant that if you had no logins but did have bookmarks, you couldn't import.
When you have never saved a login, the importer was treating it as a failed import. Now, if you have a key3/4.db file but no logins.json, it will treat it as a successful import of 0 logins.
The profile code was also not factoring in the browser it was attached to, so if you had a Mozilla profile with no logins but it somehow had a file named "Login Data" (which Chrome uses), it would think it had login data anyway. Now this code checks only the files for the browser it was created for.