Delay modal cancellation handler on setting (#952)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1203872613153566/f
Description:
There is a crash that is happening when a JS alert attempts to be displayed while another WebKit dialog is displayed (e.g printing). This likely happens when any dialog attempts to be displayed while another is already showing. There's timing issue where a completion handler which nils the property on deinit of the old dialog is called just after the new dialog is set. To prevent this, this PR dispatches the completion handler async, adding it to the next run loop, preventing the timing issue.