Update the feedback URL to open in the current window. (#17)
Task/Issue URL: https://app.asana.com/0/1203512625915051/1203889236476641/f
Tech Design URL:
CC:

Description:

This PR updates the NetP feedback link to open the link in a new tab on the current window. Previously, it was being opened in the default browser, which is not necessarily going to be the NetP client.