Present an alert to users when running the app outside of the Applications directory. (#18)
Task/Issue URL: https://app.asana.com/0/1203512625915051/1203897235340654/f
Tech Design URL:
CC:

Description:

System Extensions need their host app to run inside the /Applications directory. This is not at all obvious to users, so this PR makes the app present a warning when it is built for system extensions but not within that directory.