Sparkle support (#21)
Task/Issue URL: https://app.asana.com/0/1203137811378537/1203860366664470/f
Tech Design URL:
CC:

Description:

This PR updates the Sparkle URL to point to the NetP appcast.

Note that I began updating the archive script to work for NetP as a part of this PR, but it isn't running yet. I don't want to burn a bunch of time on it, so I'll revisit it later.