Avoid crashing when no browsers are available for data import (#978)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1203981356185148/f
Tech Design URL:
CC:

Description:

This PR fixes a rare, but apparently possible, crash when opening the data import panel. If the user has zero browsers installed, it will crash due to a force unwrap. This PR improves reliability, and introduces a new type of assertion failure which fires a pixel to let us detect when assertions are happening in the real world.