Fix occasional input freezing in pinned tabs (#979)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1202668155654614/f

Description:
Update the logic for publishing events to webViewDidReceiveChallengePublisher. Previously
it used to be fired on every webView:didReceiveChallenge: callback. This callback is called multiple
times on some websites, most importantly Asana and GitHub, also outside of regular navigation,
with authentication method transparent to the user (such as NSURLAuthenticationMethodServerTrust).

Our use of this event is to handle delayed website loading (to mimic Safari and remove white background
flashing when loading SERP while in dark mode), and our only need for handling a challenge event here is
to support presenting HTTP Basic Auth dialog to the user.

Hence, the change is to only publish an event there when didReceiveChallenge cause flow interruption
(which, for now, is only the Basic Auth scenario). Because of that and to make the API clear, the publisher
has been renamed as webViewDidReceiveUserInteractiveChallengePublisher.

Another two fixes included in this change are:

* don't call showTabContent when tab displays a URL (in this case webview stays in BrowserTabVC
  and url change is handled by the webview itself),
* always make webview first responder when switching between windows that display the same
  pinned tab (i.e. when webview snapshot is being removed).