[navigation #8] Downloads Tab Extension (#893)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.
-->

Task/Issue URL:
https://app.asana.com/0/1199230911884351/1203902804957453
CC: @tomasstrba @ayoy @Bunn 

**Description**:
- Downloads handled in DownloadsTabExtension
- includes:
  - https://github.com/more-duckduckgo-org/macos-browser/pull/988
  - https://github.com/duckduckgo/macos-browser/pull/1011
  - https://github.com/duckduckgo/macos-browser/pull/1014

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

**Steps to test this PR**:
1. Validate downloads, blob:// downloads (e.g. in whatsapp), Save Link
location As, Save Image As working as expected
2. Validate Alerts and Save Dialogs working as expected
3. Open download link in a new tab, wait for download to start, close
the tab and press cmd+shift+t: validate download is not restarted
4. Open download link in a new tab, wait for download to start, quit the
app and reopen it restoring session, validate download is not restarted;
Repeat with deactivated (background lazy loaded tab)
5. Validate download is restarted on Enter pressed in the Address Bar or
on Refresh

Test links:
- whatsapp chat attachments for blob: downloads
- https://en.wikipedia.org/wiki/Guitar#/media/File:GuitareClassique5.png
for closing tab after download starts
- https://www.openoffice.org/download/ - download full install for new
tab download and testing resume