[navigation #6] HTTPSUpgrade, PrivacyDashboard Tab Extensions (#890)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.
-->

❗️included TabContent changes made as a separate PR:
https://github.com/more-duckduckgo-org/macos-browser/pull/987

Task/Issue URL: https://app.asana.com/0/0/1203487090719126/f
Tech Design URL:https://app.asana.com/0/481882893211075/1203268245242140
CC: @tomasstrba @ayoy 

**Description**:
Moves Privacy Dashboard and HTTPSUPgrade to Tab Extensions

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

**Steps to test this PR**:
1. Validate HTTPS Upgrade scenarios work
2. Validate Privacy Dashboard is correctly handling loading states and
tracker blocked (regular, third party, surrogates)