[navigation #9] link protection (#894)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.
-->

Task/Issue: https://app.asana.com/0/1199230911884351/1203902804957455/f
Discussion: https://app.asana.com/0/1202406491309510/1203538050625396
CC: @ayoy @tomasstrba @Bunn 

**Description**:
- Tracking Link Rewriting, Referrer trimming, GPC moved to a
TabExtension
- Organised what‘s left in the Tab+Navigation file and moved away
TabContent
- Depends on #893 

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

**Steps to test this PR**:
1. Validate privacy protections work:
(https://privacy-test-pages.glitch.me/)
- Link Protection (AMP Links Protection:
getCleanURL/trackingLinkRewrite)
  - Referrer Trimming
  - Global Privacy Control (GPC)
2. Validate other browser parts work
  - opening link in another tab/window, privacy protections should apply
  - toggling protections on/off