[navigation #10] Tab History + DuckPlayer + Tab hotkeys handler (#898)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.
-->

Task/Issue: [Tab Navigation
stack](https://app.asana.com/0/0/1204014501372127/f)
CC: @ayoy 

**Description**:
- Tab History moved to Tab Extension
- Adjusted Tab Content for future Content Flow changes but still a
temporary solution so standardise Tab exported publishers

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

**Steps to test this PR**:
1. Validate History works as intended
2. Validate Tab content/title change is handled correctly
(navigation/no-navigation like asana links: reflects title change and
url, setting tab.content updates title/url and causes navigation)

**Testing checklist**:

* [ ] Test with Release configuration