[navigation #5] content blocking scripts, click to load TabExtensions (#888)
<!--
Note: This checklist is a reminder of our shared engineering
expectations. Feel free to change it, although assigning a GitHub
reviewer and the items in bold are required.
-->

Task/Issue URL: https://app.asana.com/0/0/1203487090719126/f
Tech Design URL:
https://app.asana.com/0/481882893211075/1203268245242140/f
CC: @tomasstrba @bwaresiak 

**Description**:
FBProtection(click to load) and ContentBlockingUserScripts/surrogates
handling moved to Tab Extensions

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

**Steps to test this PR**:
1. Validate Click To Load works
https://privacy-test-pages.glitch.me/privacy-protections/click-to-load/
2. Validate Privacy Dashboard is updated with regular, 3rd party and
surrogate blocked trackers
3. Ensure awaitContentBlockingAssetsInstalled is called on non-ddg
navigations