Fix bookmarks popover folder renaming (#1046)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1204209081295753/f
Tech Design URL:
CC:

Description:

This PR fixes an issue where folders couldn't be renamed from the bookmarks popover. The issue was two-fold:

The delegate wasn't being set at all
Even if it were, the delegate is the popover which was being deallocated after opening the rename folder dialog
To fix this, we are now setting the delegate and using the shouldClosePopover functionality to keep the popover alive while the alert is active. This is a better UX than before, as the user gets to stay in the popover while they edit, instead of having to reopen it again once they're done.