Add active usage pixels, and support daily/continuous pixel formats (#43)
Task/Issue URL: https://app.asana.com/0/0/1203934440599379/f
Tech Design URL:
CC:

Description:

This PR adds support for daily usage pixels, which are triggered as a part of the rekeying timer.

It also adds support for daily/continuous pixels, including maintaining support for the old format. This is now specified as a non-optional pixel parameter, to force us to be explicit about how pixels are being sent.