Update backend URL (#45)
Task/Issue URL: https://app.asana.com/0/1203512625915051/1204259203255704/f
Tech Design URL:
CC:

Description:

This PR updates the NetP backend endpoint. There are no changes to the response from the endpoint.

I added a new test file which is a copy of the current servers list from the new endpoint to be on the safe side.