Sabrina/global zoom setting (#1056)
Task/Issue URL:
https://app.asana.com/0/72649045549333/1203353307511501/f

Introduce a global zoom setting to allow users changing
the default zoom:
In the appearance settings a new zoom setting ha been added. It is
possible to select a zoom level which will be applied to all WebView
The allowed predefined levels will be:  50, 75, 85, 100, 115, 125, 150,
175, 200, 250, 300%.
A new "Change default page zoom..." item has been added to the more
option menu. Clicking on this item takes the user to the appearance
setting page.
The same stops will be used for the Default Zoom and for the Zoom In/Out
functions on the page.
The actual zoom size in the zoom more options menu should be adapted to
the default zoom level