YouTube UserScript as proxy for SERP (#959)
Task/Issue URL: https://app.asana.com/0/0/1203884586225701/f
Tech Design URL:
CC: @ayoy @brianhall 

**Description**:

Tweaks the YouTube Overlay UserScript to act as a proxy between page &
user-scripts. This enables the SERP to interact with settings WITHOUT
direct access to any of the webkit messaging handlers 💪🏻

This also adds 3 new pixels

**Steps to test this PR**:
1. @brianhall will conduct testing on the SERP side, details in the
task2.
2. I'm not sure on the process for E2E testing the pixels?

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
**When ready for review, remember to post the PR in MM**

---------

Co-authored-by: Shane Osbourne <sosbourne@duckduckgo.com>