Add a common release workflow for making DMG + App Store releases (#1117)
Task/Issue URL: https://app.asana.com/0/1203301625297703/1204375817149221/f

Description:
The "Prepare Release" workflow is supposed to be the one workflow that's supposed to be
run to have DMG and App Store builds prepared while making a release. It takes a single argument
– the Asana task URL (for DMG/dSYM.zip upload) and sets all other parameters itself.
DMG and App Store builds are run in parallel.