Fix potential render/selection/scroll issue in Autofill List (#1130)
Task/Issue URL: https://app.asana.com/0/1204099484721401/1204318806095433/f

Description:
In some cases, opening the Autofill popover causes the scroll to be janky or locked (see this video).

This issue has been around for a long time. We've tried fixing without luck here without luck.

It's still unclear why this is happening, but it may be related to the lazy stack rendering..

A. The view takes too long to render when you have several hundreds of login items, causing the selection/scroll to have issues.
B. The model not returning the list of items before we trigger the selection, which happens onAppear.

I've implemented a few changes and moved the selection to .onChange, adding a dynamic delay based on the list of items in the user's vault, so the list has more time to render with bigger datasets.

Still, not an ideal solution until Apple implements a proper .onLayout method for SwiftUI views.