Organize the project file (#66)
Task/Issue URL: https://app.asana.com/0/1203512625915051/1204474500868594/f
Tech Design URL:
CC:

Description:

This PR organizes the NetP files to make the upstream merge easier. As a part of that, I'm going to have to re-apply project file changes by hand, and it's not clear from the file system as to which files go into which targets - this PR attempts to address that, by organizing files by their target.

Let me know if you think the organization could use work, I'm open to suggestions – I implemented this with the goal of making my life easier during the merge, we can always change it before or after that as it will no longer be as important to maintain a structure like this.

I have also updated the apps to all use the same Info.plist file, as we prepare to move NetP into the main repo.