Check that the source and target frame security origins are equal (#1207)
Task/Issue URL: https://app.asana.com/0/1176047645786601/1202592488248250/f
Tech Design URL:
CC: @mallexxx

Description:

This PR checks that the security origins of the target and source frames are equivalent. This is to prevent a situation where one frame is requesting an external URL request for a second frame, which makes it possible for a permission prompt to appear in a tab that didn't request it.